requirements.txt
alembic==1.13.0
Babel==2.14.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-mailman==1.0.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.6
importlib-resources==6.1.1
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
passlib==1.7.4
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
tzdata==2023.4
validators==0.22.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
xlrd==2.0.1
